refactor: move use-render-props and use-editor-props to lib/props
