+++ packages/core/components/DropZone/index.tsx
@@ -47,8 +47,8 @@ import { SlotRenderPure } from "../SlotRender";
 import { expandNode } from "../../lib/data/flatten-node";
 import { useMessage } from "../../lib/use-message";
 import { FieldTransforms } from "../../types/API/FieldTransforms";
-import useEditorProps from "../../lib/field-transforms/use-editor-props";
-import useRenderProps from "../../lib/field-transforms/use-render-props";
+import useEditorProps from "../../lib/props/editor/use-editor-props";
+import useRenderProps from "../../lib/props/render/use-render-props";
 import { MemoizeComponent } from "../MemoizeComponent";
 import { VirtualizedDropZone } from "./VirtualizedDropZone";
 import { LinePlaceholder } from "./LinePlaceholder";
