+++ packages/core/components/ServerRender/index.tsx
@@ -5,7 +5,7 @@ import {
   rootDroppableId,
   rootZone,
 } from "../../lib/root-droppable-id";
-import useRenderProps from "../../lib/field-transforms/use-render-props";
+import useRenderProps from "../../lib/props/render/use-render-props";
 import { setupZone } from "../../lib/data/setup-zone";
 import {
   Config,
